+++ docsig/_cli.py
@@ -2,24 +2,56 @@
 docsig._cli
 ===========
 """
+import re as _re
 import sys as _sys
+import typing as _t
 from argparse import ArgumentParser as _ArgumentParser
+from argparse import HelpFormatter as _HelpFormatter
 from pathlib import Path as _Path
 
+import tomli as _tomli
+
 from ._utils import color as _color
+from ._utils import find_pyproject_toml as _find_pyproject_toml
 from ._version import __version__
 
 NAME = __name__.split(".", maxsplit=1)[0]
+
+ConfigType = _t.Dict[str, _t.Any]
+
+
+# split str by comma, but allow for escaping
+def _split_comma(value: str) -> _t.List[str]:
+    return [i.replace("\\,", ",") for i in _re.split(r"(?<!\\),", value)]
+
+
+def get_config() -> ConfigType:
+    """Get config dict object from package's tool section in toml file.
+
+    :return: Dict object; containing config if there is one, else return
+        empty.
+    """
+    pyproject_file = _find_pyproject_toml()
+    if pyproject_file is None:
+        return {}
+
+    return (
+        _tomli.loads(pyproject_file.read_text()).get("tool", {}).get(NAME, {})
+    )
 
 
 class Parser(_ArgumentParser):
     """Parse commandline arguments."""
 
-    def __init__(self) -> None:
+    def __init__(self, kwargs: ConfigType) -> None:
         super().__init__(
             prog=_color.cyan.get(NAME),
+            formatter_class=lambda prog: _HelpFormatter(
+                prog, max_help_position=45
+            ),
             description="Check docstring matches signature",
         )
+        self._kwargs = kwargs
         self._add_arguments()
         self._version_request()
         self.args = self.parse_args()
@@ -37,6 +69,15 @@ class Parser(_ArgumentParser):
             "--version",
             action="store_true",
             help="show version and exit",
+        )
+        self.add_argument(
+            "-d",
+            "--disable",
+            action="store",
+            metavar="LIST",
+            type=_split_comma,
+            default=self._kwargs.get("disable", []),
+            help="comma separated list of rules to disable",
         )
 
     @staticmethod